refactor(assets): one total band for the dashboard panel and the page
The portfolio total was a private widget inside coins_screen.dart, which was
fine while the dashboard panel was the only surface printing it. The /assets
page is about to print the same number, and two copies of a number the user
reads as one fact drift: the type, the colour, or the optical centring of the
24h percentage would sooner or later disagree between the two screens.

Promoted verbatim - same tokens, same space4 inset, same
CrossAxisAlignment.center that fixed the percentage hanging low. The
dashboard's rendered tree is unchanged, and the header test proves it: the
same boxH 32.0, drop 0.125 and header 94.0 it measured before.

The test's finder moved with it. It matched the private type by name, and the
test file declares its OWN AssetsTotalBand as a measurement probe, so the
shipping widget is imported with a prefix rather than either being renamed - a
probe called something else drifts out of step with what it claims to mirror.
